chore: refactor componenets
