+++ src/App.tsx
@@ -1,5 +1,4 @@
 import Navigation from "./Navigation";
-import { GlobalErrorProvider } from "./providers/GlobalError";
 import { WagmiProvider } from "./providers/Wagmi";
 import { WalletModalProvider } from "./providers/WalletModal";
 import ApolloProvider from "./providers/Apollo";
@@ -8,15 +7,13 @@ import { ConfigProvider } from "./providers/Config";
 const App = () => {
   return (
     <ConfigProvider>
-      <GlobalErrorProvider>
-        <WagmiProvider>
-          <WalletModalProvider>
-            <ApolloProvider>
-              <Navigation />
-            </ApolloProvider>
-          </WalletModalProvider>
-        </WagmiProvider>
-      </GlobalErrorProvider>
+      <WagmiProvider>
+        <WalletModalProvider>
+          <ApolloProvider>
+            <Navigation />
+          </ApolloProvider>
+        </WalletModalProvider>
+      </WagmiProvider>
     </ConfigProvider>
   );
 };
