+++ src/components/Buttons/ConnectWalletButton.tsx
@@ -1,12 +1,12 @@
 import React from "react";
-import Button from "../Button/Button_View";
 import { useAccount, useDisconnect } from "wagmi";
+import { BaseButton } from ".";
 
 type Props = React.ButtonHTMLAttributes<HTMLButtonElement> & {
   openWalletModal: () => void;
 };
 
-const WalletConnectButton: React.FC<Props> = ({ openWalletModal }) => {
+export const ConnectWalletButton: React.FC<Props> = ({ openWalletModal }) => {
   const { address, isConnected, isConnecting } = useAccount();
   const { disconnect } = useDisconnect();
 
@@ -23,22 +23,22 @@ const WalletConnectButton: React.FC<Props> = ({ openWalletModal }) => {
           <div className="inline-flex items-center mx-2 gap-4">
             <span className="truncate text-gray-500">{address}</span>
             <div>
-              <Button
+              <BaseButton
                 className="cursor-pointer hover:bg-gray-200 hover:text-white sm:w-auto sm:mb-0"
                 onClick={() => disconnect()}
               >
                 Disconnect
-              </Button>
+              </BaseButton>
             </div>
           </div>
         ) : (
           <div className="mx-4">
-            <Button onClick={openWalletModal}>Connect your Wallet</Button>
+            <BaseButton onClick={openWalletModal}>
+              Connect your Wallet
+            </BaseButton>
           </div>
         )}
       </div>
     </>
   );
 };
-
-export default WalletConnectButton;
