+++ src/components/Buttons/index.ts
@@ -1,0 +1,2 @@
+export * from "./BaseButton";
+export * from "./ConnectWalletButton";
