+++ src/components/Loader.tsx
@@ -5,14 +5,6 @@ type Props = {
   size?: number;
 };
 
-const Loader: React.FC<Props> = ({ className, size = 30 }) => {
-  return (
-    <ClipLoader
-      aria-label="Loading Spinner"
-      className={className}
-      size={size}
-    />
-  );
-};
-
-export default Loader;
+export const Loader: React.FC<Props> = ({ className, size = 30 }) => (
+  <ClipLoader aria-label="Loading Spinner" className={className} size={size} />
+);
