+++ src/components/Modals/WalletModal.tsx
@@ -1,0 +1,62 @@
+import React, { useEffect } from "react";
+import { useAccount, useConnect } from "wagmi";
+import { MetaMaskIcon, WalletConnectIcon } from "../../icons";
+import { BaseModal } from ".";
+
+type Props = {
+  isModalOpen: boolean;
+  closeWalletModal: () => void;
+};
+
+export const WalletModal: React.FC<Props> = (props: Props) => {
+  const { isModalOpen, closeWalletModal } = props;
+  const { isConnected } = useAccount();
+  const { connect, connectors } = useConnect();
+
+  useEffect(() => {
+    if (isConnected) {
+      closeWalletModal();
+    }
+  }, [isConnected, closeWalletModal]);
+
+  return (
+    <BaseModal isOpen={isModalOpen} onClose={closeWalletModal}>
+      <div className="text-center w-96">
+        <div>
+          <label
+            className="flex justify-center cursor-pointer"
+            onClick={e => {
+              e.preventDefault();
+              connect({ connector: connectors[0] });
+            }}
+          >
+            <MetaMaskIcon
+              title="meta-mask-icon"
+              className="w-20 h-20 transition-opacity duration-500 ease-out opacity-100 hover:opacity-40"
+            />
+          </label>
+          <div className="font-bold">METAMASK</div>
+          <div>Connect using your browser.</div>
+        </div>
+
+        <div className="py-4 mb-3 border-0 border-b border-solid"></div>
+        <div>
+          <label
+            className="flex justify-center cursor-pointer"
+            onClick={e => {
+              e.preventDefault();
+              connect({ connector: connectors[1] });
+            }}
+          >
+            <WalletConnectIcon
+              title="wallet-connect-icon"
+              className="w-20 h-20 transition-opacity duration-500 ease-out opacity-100 hover:opacity-40"
+            />
+          </label>
+          <div className="font-bold">WALLET CONNECT</div>
+          <div>Connect using your mobile device.</div>
+        </div>
+      </div>
+    </BaseModal>
+  );
+};
