+++ src/components/Modals/index.ts
@@ -2,3 +2,6 @@ export * from "./PlaceBetModal";
 export * from "./SettleBetModal";
 export * from "./DepositVaultModal";
 export * from "./WithdrawVaultModal";
+export * from "./WalletModal";
+export * from "./BaseModal";
+export * from "./FaucetModal";
