--- src/components/PageLayout/PageLayout_Logic.tsx
@@ -1,18 +1,0 @@
-import { useLocation } from "react-router";
-import PageLayoutView from "./PageLayout_View";
-
-type Props = {
-  children: React.ReactNode;
-};
-
-const PageLayout: React.FC<Props> = props => {
-  const location = useLocation();
-
-  return (
-    <PageLayoutView loading={false} currentPath={location.pathname}>
-      {props.children}
-    </PageLayoutView>
-  );
-};
-
-export default PageLayout;
