+++ src/components/Races/index.ts
@@ -1,0 +1,2 @@
+export * from "./RaceTable";
+export * from "./RaceTableRow";
