+++ src/components/Toggle.tsx
@@ -5,22 +5,18 @@ type Props = {
   onChange: () => void;
 };
 
-const Toggle = ({ enabled, onChange }: Props) => {
-  return (
-    <Switch
-      checked={enabled}
-      onChange={onChange}
-      className={`${enabled ? "bg-blue-600" : "bg-gray-200"}
-            relative inline-flex h-6 w-11 items-center rounded-full`}
-    >
-      <span className="sr-only">Use setting</span>
-      <span
-        aria-hidden="true"
-        className={`${enabled ? "translate-x-6" : "translate-x-1"}
-              inline-block h-4 w-4 transform rounded-full bg-white transition`}
-      />
-    </Switch>
-  );
-};
-
-export default Toggle;
+export const Toggle: React.FC<Props> = ({ enabled, onChange }) => (
+  <Switch
+    checked={enabled}
+    onChange={onChange}
+    className={`${enabled ? "bg-blue-600" : "bg-gray-200"}
+          relative inline-flex h-6 w-11 items-center rounded-full`}
+  >
+    <span className="sr-only">Use setting</span>
+    <span
+      aria-hidden="true"
+      className={`${enabled ? "translate-x-6" : "translate-x-1"}
+            inline-block h-4 w-4 transform rounded-full bg-white transition`}
+    />
+  </Switch>
+);
