+++ src/components/Vaults/VaultHistoryRow.tsx
@@ -6,19 +6,20 @@ import { Config } from "../../types/config";
 import utils from "../../utils";
 import { VaultTransactionType } from "../../types/vaults";
 
-type Props = {
-  vault: FormattedVaultTransaction;
-  config?: Config;
-};
-
 const txTypeMap = new Map([
   [VaultTransactionType.WITHDRAW, "Withdrawal"],
   [VaultTransactionType.DEPOSIT, "Deposit"]
 ]);
 
+type Props = {
+  vault: FormattedVaultTransaction;
+  config?: Config;
+};
+
 export const VaultHistoryRow: React.FC<Props> = ({ vault, config }) => {
   const formattedTxType = txTypeMap.get(vault.type);
   const details = utils.config.getVault(vault.vaultAddress, config);
+
   return (
     <tr>
       <td className="pl-5 pr-2 py-4 whitespace-nowrap">{formattedTxType}</td>
