+++ src/components/index.ts
@@ -1,0 +1,5 @@
+export * from "./Transition";
+export * from "./Toggle";
+export * from "./PageLayout";
+export * from "./Loader";
+export * from "./Card";
