+++ src/pages/Tokens/Tokens_View.tsx
@@ -1,15 +1,13 @@
-import PageLayout from "../../components/PageLayout/PageLayout_Logic";
+import { PageLayout } from "../../components";
 
-const TokensView: React.FC = () => {
-  return (
-    <PageLayout>
-      <div className="flex flex-col">
-        <div className="-my-2 overflow-x-auto sm:-mx-6 lg:-mx-8">
-          <h1>Tokens</h1>
-        </div>
+const TokensView: React.FC = () => (
+  <PageLayout>
+    <div className="flex flex-col">
+      <div className="-my-2 overflow-x-auto sm:-mx-6 lg:-mx-8">
+        <h1>Tokens</h1>
       </div>
-    </PageLayout>
-  );
-};
+    </div>
+  </PageLayout>
+);
 
 export default TokensView;
